Fix #454
