+++ src/utils/siteSections.ts
@@ -1,0 +1,10 @@
+// NYPL site sections. See Figma: https://www.figma.com/file/qShodlfNCJHb8n03IFyApM/Main?node-id=15454%3A47007
+
+const sections = [
+  "nypl--books-and-more",
+  "nypl--locations",
+  "nypl--research",
+  "nypl--whats-on",
+];
+
+export default sections;
